pyproject.toml
[project]
name = "gamedeals"
version = "0.1.0"
description = "Scours Reddit's r/GameDeals for free game offers and emails them to you."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pystache>=0.6.8",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
